expand module, + some code cleaning
